awful.mouse: Move snap into a submodule
